% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPProbit-package.R
\docType{package}
\name{BayesPProbit}
\alias{BayesPProbit}
\title{Bayesian p-Generalized Probit Regression}
\description{
Implements Bayesian inference for p-generalized probit regression models with
efficient data compression using coresets. The package provides tools for:
\itemize{
\item Fitting Bayesian p-generalized probit models
\item Constructing coresets for scalable computation
\item Posterior inference and diagnostics
\item Visualization of results
}
}
\section{Main Functions}{

\describe{
\item{multi_chain}{Run MCMC sampling with multiple chains}
\item{compute_coreset}{Construct coresets for data compression}
\item{plot_diagnostics}{Plot MCMC diagnostics}
}
}

\references{
Ding et al. (2024). Scalable Bayesian p-generalized probit and logistic
regression. Advances in Data Analysis and Classification.
}
