% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{compute_leverage_scores}
\alias{compute_leverage_scores}
\title{Compute Leverage Scores}
\usage{
compute_leverage_scores(X, p, sketch_size)
}
\arguments{
\item{X}{Design matrix}

\item{p}{Shape parameter}

\item{sketch_size}{Size of sketching matrix}
}
\value{
Vector of leverage scores for each observation
}
\description{
Computes p-norm leverage scores using sketching technique
}
\keyword{internal}
