% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{compute_oneshot_scores}
\alias{compute_oneshot_scores}
\title{Compute One-shot Coreset Scores}
\usage{
compute_oneshot_scores(X, p_max, sketch_size)
}
\arguments{
\item{X}{Design matrix}

\item{p_max}{Maximum p value to consider}

\item{sketch_size}{Size of sketching matrix}
}
\value{
Vector of sensitivity scores
}
\description{
Computes sampling scores valid for multiple p values
}
\keyword{internal}
