% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{compute_sketch}
\alias{compute_sketch}
\title{Compute Matrix Sketch}
\usage{
compute_sketch(X, sketch_size, p)
}
\arguments{
\item{X}{Input matrix}

\item{sketch_size}{Size of sketching matrix}

\item{p}{Shape parameter}
}
\value{
Sketched matrix
}
\description{
Computes Count-Sketch or Gaussian sketch of input matrix
}
\keyword{internal}
