% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs Sampler Implementation}
\usage{
gibbs_sampler(
  n_sim,
  burn_in,
  X,
  y,
  initial_theta,
  initial_p,
  mh_iter,
  p_range,
  step_size,
  fix_p = FALSE
)
}
\arguments{
\item{n_sim}{Number of iterations}

\item{burn_in}{Burn-in period}

\item{X}{Design matrix}

\item{y}{Response vector}

\item{initial_theta}{Initial parameters}

\item{initial_p}{Initial p value}

\item{mh_iter}{MH iterations}

\item{p_range}{Range for p}

\item{step_size}{Step size}
}
\value{
List containing:
\describe{
\item{beta_chain}{MCMC chain for beta parameters}
\item{p_chain}{Chain for p parameter}
}
}
\description{
Internal implementation of Gibbs sampler for p-generalized probit model
}
\keyword{internal}
