% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log Likelihood for p-Probit Model}
\usage{
log_likelihood(X, beta, p, y)
}
\description{
Computes log likelihood for p-generalized probit regression
}
\keyword{internal}
