% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_scale}
\alias{p_scale}
\title{Scale Parameter for p-GGD}
\usage{
p_scale(p)
}
\arguments{
\item{p}{Shape parameter}
}
\value{
A numeric value representing the scale parameter p^(1/p)
}
\description{
Computes scale parameter for p-generalized Gaussian distribution
}
