% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pggd}
\alias{pggd}
\alias{dgnorm}
\alias{pgnorm}
\alias{qgnorm}
\alias{rgnorm}
\title{p-Generalized Gaussian Distribution Functions}
\usage{
dgnorm(x, mu = 0, alpha = 1, beta = 2)

pgnorm(x, mu = 0, alpha = 1, beta = 2)

qgnorm(p, mu = 0, alpha = 1, beta = 2)

rgnorm(n, mu = 0, alpha = 1, beta = 2)
}
\description{
Density, distribution, quantile and random generation functions
for the p-generalized Gaussian distribution
}
