% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{plot.pgprobit}
\alias{plot.pgprobit}
\title{Plot Method for pgprobit Objects}
\usage{
\method{plot}{pgprobit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pgprobit}}

\item{...}{Additional arguments passed to or from other methods}
}
\value{
No return value, called for side effects of creating diagnostic plots.
}
\description{
Creates diagnostic plots for Bayesian p-Generalized Probit Regression results.
}
