% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{print.pgprobit}
\alias{print.pgprobit}
\title{Print Method for pgprobit Objects}
\usage{
\method{print}{pgprobit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pgprobit}}

\item{...}{Additional arguments passed to or from other methods}
}
\value{
No return value, called for side effects of printing summary information.
}
\description{
Prints a summary of the Bayesian p-Generalized Probit Regression results.
}
