% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtrunc_pggd}
\alias{rtrunc_pggd}
\title{Sample from Truncated p-GGD}
\usage{
rtrunc_pggd(n, mu, p, lower = -Inf, upper = Inf)
}
\description{
Generates samples from truncated p-generalized Gaussian distribution
}
\keyword{internal}
