% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{seq_geom}
\alias{seq_geom}
\title{Generate Geometric Sequence}
\usage{
seq_geom(from, to, by)
}
\arguments{
\item{from}{Starting value}

\item{to}{Ending value}

\item{by}{Geometric step size}
}
\value{
Vector of geometrically spaced values
}
\description{
Generates sequence with geometric spacing
}
\keyword{internal}
