% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{compute_oneshot_scores}
\alias{compute_oneshot_scores}
\title{Compute One-shot Coreset Scores}
\usage{
compute_oneshot_scores(X, p_max, sketch_size)
}
\description{
Computes sampling scores valid for multiple p values
}
\keyword{internal}
