% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs Sampler Implementation}
\usage{
gibbs_sampler(
  n_sim,
  burn_in,
  X,
  y,
  initial_theta,
  p,
  mh_iter,
  p_range,
  step_size
)
}
\description{
Gibbs Sampler Implementation
}
\keyword{internal}
